# swagger.yml
swagger: "2.0"
info:
  description: API для работы с транзакциями
  version: "1.0.0"
  title: Транзакции API
paths:
...
  /transactions/{transaction_id}:
    get:
      summary: Получение детальной информации о транзакции
      description: Возвращает полную информацию о транзакции, включая статус и цвет.
      parameters:
        - name: transaction_id
          in: path
          required: true
          type: integer
          description: ID транзакции.
      responses:
        200:
          description: Информация о транзакции
          schema:
            type: object
            properties:
              id:
                type: integer
                example: 123
              amount:
                type: number
                format: float
                example: 100.0
              status:
                type: string
                example: "confirmed"
              created_at:
                type: string
                format: date-time
                example: "2025-01-12T10:30:00"
        404:
          description: Транзакция не найдена
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Transaction not found."